286. SpawnActorDeferred and BeginPlay (SPAWNPOINT BUGFIX!)
The previous way of spawning wasn't correct, since the SpawnActor code, also actioned the beginplay() code of the spawning actor, before we got to the next line of attaching the newly spawned actor, to a component.
A way to post-pone the BeginPlay() code in the spawned actor is to use UWorld::SpawnActorDeferred(), followed by UGameplayStatics::FinishSpawningActor(), to initiate it's begin play.
- https://docs.unrealengine.com/en-US/API/Runtime/Engine/Engine/UWorld/SpawnActorDeferred/index.html
- https://docs.unrealengine.com/en-US/API/Runtime/Engine/Kismet/UGameplayStatics/FinishSpawningActor/index.html

Changes:
- Replaced the spawning code in Spawnpoint, top use SpawnActorDeferred() and FinishSpawningActor() instead.
- Reset the location of the spawnpoint in the blueprint (it wasn't set to 0 properly, in relative space).
- Used 'GetRelativeTransform()' instead of the lecture's suggestion to use 'GetComponentTransform(), and Change the AttachToComponent arguement to "FAttachmentTransformRules::KeepWorldTransform", instead of 'Relative'.